% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.R
\name{acc_sandwich_hermite}
\alias{acc_sandwich_hermite}
\title{Sandwich Estimator for Augmented Complete Case for Censored Covariates}
\usage{
acc_sandwich_hermite(
  formula,
  data,
  par_vec,
  cens_name,
  cov_vars,
  beta_est,
  cens_ind,
  cov_dist_params,
  sigma2,
  cov_dist_opt,
  gh_nodes
)
}
\arguments{
\item{formula}{a linear or nonlinear model formula including variables and parameters}

\item{data}{a data frame containing columns for the censoring indicator and the variables in \code{formula}}

\item{par_vec}{a character string indicating the parameter vector in the formula}

\item{cens_name}{a character string indicating the name of censored covariate from \code{data}}

\item{cov_vars}{if \code{cov_dist_opt} one of \code{c("MVN")}, a list of character strings indicating the names of the variables from \code{data} to be used as predictors in the covariate distribution Otherwise \code{NULL}.}

\item{beta_est}{the estimate from the augmented complete case estimator}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{cov_dist_params}{a list of parameters for the conditional distribution of the censored covariate given the outcome and fully observed covariates}

\item{sigma2}{the estimate of the error variance}

\item{cov_dist_opt}{a character string indicating specification of the covariate distribution. One of "MVN", "user_MVN", "AFT"}

\item{gh_nodes}{number of nodes to use in gauss-hermite quadrature.}
}
\value{
A vector of the sandwich standard error estimates.
}
\description{
Calculates the standard error sandwich estimate for an augmented complete case estimator for a regression model with censored covariates.
}
