% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{aipw_censored}
\alias{aipw_censored}
\title{Augmented Inverse Probability Weighting for Censored Covariates}
\usage{
aipw_censored(
  formula,
  data,
  cens_ind,
  cens_name,
  par_vec,
  starting_vals,
  sandwich_se = TRUE,
  weight_opt,
  weights_user = NULL,
  weights_cov = NULL,
  weights_threshold = NULL,
  weight_stabilize = "None",
  cov_dist_opt = "MVN",
  cov_vars,
  cov_mean_user = NULL,
  cov_sigma_user = NULL,
  gh = TRUE,
  gh_nodes = 10,
  ...
)
}
\arguments{
\item{formula}{a linear or nonlinear model formula including variables and parameters}

\item{data}{a data frame containing columns for the censoring indicator and the variables in the formula}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{cens_name}{a character string indicating the name of censored covariate from \code{data}}

\item{par_vec}{a character string indicating the parameter vector in the formula}

\item{starting_vals}{the starting values for the least squares algorithm. Must be a vector equal in length of the parameter vector}

\item{sandwich_se}{if \code{TRUE} (default), the empirical sandwich estimator for the standard error is calculated}

\item{weight_opt}{a character string indicating the method of weight calculation. One of "Cox", "AFT_lognormal", "MVN", "user" (if "user", then user provides weights).}

\item{weights_user}{if \code{weight_opt = "user"}, a vector of weights the same length as there are rows in \code{data}, otherwise \code{NULL} (default).}

\item{weights_cov}{if \code{weight_opt} one of \code{c("Cox", "AFT_lognormal", "MVN")}, a list of character strings indicating the names of the variables from \code{data} to be used as predictors in the weights model. Otherwise \code{NULL}.}

\item{weights_threshold}{the maximum weight for any one observation. If \code{NULL} (default), there is no thresholding.}

\item{weight_stabilize}{a character string indicating which method of weight stabilization is to be done (if any). One of \code{c("Mean", "KM", "None")}.}

\item{cov_dist_opt}{a character string indicating specification of the covariate distribution. One of "MVN", "user MVN", "AFT"}

\item{cov_vars}{if \code{cov_dist_opt} one of \code{c("MVN")}, a list of character strings indicating the names of the variables from \code{data} to be used as predictors in the covariate distribution Otherwise \code{NULL}.}

\item{cov_mean_user}{if \code{cov_dis_opt = "user MVN"}, the mean of the multivariate normal distribution of \code{(log(X), log(C), Z)}.}

\item{cov_sigma_user}{if \code{cov_dis_opt = "user MVN"}, the covariance matrix of the multivariate normal distribution of \code{(log(X), log(C), Z)}.}

\item{gh}{if \code{TRUE} (default), gauss-hermite quadrature will be run to estimate the integrals. Otherwise, the \code{integrate} function will be used.}

\item{gh_nodes}{number of nodes to use in gauss-hermite quadrature.}

\item{...}{additional arguments passed to function \code{multiroot}.}
}
\value{
A list with the following elements:
\item{beta_est}{a vector of the parameter estimates.}
\item{se_est}{if \code{sandwich_se = TRUE}, a vector of standard error estimates from the empirical sandwich estimator. Otherwise, \code{NULL}}
\item{iteration_count}{the number of iterations used in \code{multiroot}.}
}
\description{
Performs an augmented inverse probability weighting estimation for a regression model with censored covariates.
}
