% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc_censored}
\alias{cc_censored}
\title{Complete Case Analysis for Censored Covariates}
\usage{
cc_censored(formula, D, data, par_vec, starting_vals, sandwich_se = TRUE)
}
\arguments{
\item{formula}{a linear or nonlinear model formula including variables and parameters.}

\item{D}{name of censoring indicator, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored.}

\item{data}{a data frame containing columns for the censoring indicator and the variables in the formula}

\item{par_vec}{name of parameter vector in the formula}

\item{starting_vals}{the starting values for the least squares algorithm. Must be a vector equal in length of the parameter vector}

\item{sandwich_se}{if \code{TRUE} (default), the empirical sandwich estimator for the standard error is calculated.}
}
\value{
A list with the following elements:
\item{beta_est}{a vector of the parameter estimates.}
\item{se_est}{if \code{sandwich_se = TRUE}, a vector of standard error estimates from the empirical sandwich estimator. Otherwise, \code{NULL}}
}
\description{
Performs a complete case analysis using least squares for a regression model with censored covariates.
}
