% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc_sandwich}
\alias{cc_sandwich}
\title{Sandwich Estimator for Complete Case for Censored Covariates}
\usage{
cc_sandwich(formula, data, cens_ind, par_vec, beta_est)
}
\arguments{
\item{formula}{a linear or nonlinear model formula including variables and parameters}

\item{data}{a data frame containing columns for the censoring indicator and the variables in the formula}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{par_vec}{a character string indicating the parameter vector in the formula}

\item{beta_est}{the estimate from the complete case estimator}
}
\value{
A vector of the sandwich standard error estimates.
}
\description{
Calculates the standard error sandwich estimate for a complete case estimator for a regression model with censored covariates.
}
