% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{ipw_censored}
\alias{ipw_censored}
\title{Inverse Probability Weighting for Censored Covariates}
\usage{
ipw_censored(
  formula,
  data,
  cens_ind,
  cens_name,
  par_vec,
  starting_vals,
  sandwich_se = TRUE,
  weight_opt,
  weights_user = NULL,
  weights_cov = NULL
)
}
\arguments{
\item{formula}{a linear or nonlinear model formula including variables and parameters}

\item{data}{a data frame containing columns for the censoring indicator and the variables in the formula}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{cens_name}{a character string indicating the name of censored covariate from \code{data}}

\item{par_vec}{a character string indicating the parameter vector in the formula}

\item{starting_vals}{the starting values for the least squares algorithm. Must be a vector equal in length of the parameter vector}

\item{sandwich_se}{if \code{TRUE} (default), the empirical sandwich estimator for the standard error is calculated}

\item{weight_opt}{a character string indicating which method of weight calculation is to be done. One of "Cox", "AFT_lognormal", "MVN", "user" (if "user", then user provides weights)}

\item{weights_user}{if \code{weight_opt = "user"}, a vector of weights the same length as there are rows in \code{data}, otherwise \code{NULL} (default)}

\item{weights_cov}{if \code{weight_opt} one of \code{c("Cox", "AFT_lognormal", "MVN")}, a list of character strings indicating the names of the variables from \code{data} to be used as predictors in the weights model. Otherwise \code{NULL}}
}
\value{
A list with the following elements:
\item{beta_est}{a vector of the parameter estimates.}
\item{se_est}{if \code{sandwich_se = TRUE}, a vector of standard error estimates from the empirical sandwich estimator. Otherwise, \code{NULL}}
}
\description{
Performs an inverse probability weighting estimation using least squares for a regression model with censored covariates.
}
