% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{ipw_sandwich}
\alias{ipw_sandwich}
\title{Sandwich Estimator for Inverse Probability Weighting for Censored Covariates}
\usage{
ipw_sandwich(
  formula,
  data,
  cens_ind,
  par_vec,
  beta_est,
  weights,
  se_opt = "known",
  params = NULL
)
}
\arguments{
\item{formula}{a linear or nonlinear model formula including variables and parameters}

\item{data}{a data frame containing columns for the censoring indicator and the variables in the formula}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{par_vec}{a character string indicating the parameter vector in the formula}

\item{beta_est}{the estimate from the complete case estimator}

\item{weights}{a vector of weights for the IPW estimating function}

\item{se_opt}{a character string indicating if the weights are assumed to be known or if they are estimated. One of \code{c("known", "est_MVN")}.}

\item{params}{if \code{se_opt = "est_MVN"}, a vector of nine parameter estimates representing those estimated from the maximum likelihood technique.}
}
\value{
A vector of the sandwich standard error estimates.
}
\description{
Calculates the standard error sandwich estimate for an IPW estimator for a regression model with censored covariates.
}
