% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_censored}
\alias{mle_censored}
\title{MLE for Censored Covariates}
\usage{
mle_censored(
  formula,
  data,
  cens_ind,
  cens_name,
  par_vec,
  starting_vals,
  sandwich_se = TRUE,
  cov_dist_opt = "MVN",
  cov_vars,
  cov_mean_user = NULL,
  cov_sigma_user = NULL,
  gh = TRUE,
  gh_nodes = 10,
  ...
)
}
\arguments{
\item{formula}{a linear or nonlinear model formula including variables and parameters}

\item{data}{a data frame containing columns for the censoring indicator and the variables in the formula}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{cens_name}{a character string indicating the name of censored covariate from \code{data}}

\item{par_vec}{a character string indicating the parameter vector in the formula}

\item{starting_vals}{the starting values for the least squares algorithm. Must be a vector equal in length of the parameter vector}

\item{sandwich_se}{if \code{TRUE} (default), the empirical sandwich estimator for the standard error is calculated}

\item{cov_dist_opt}{a character string indicating which method of covariate distribution is to be done. One of "MVN", "user MVN", "AFT_lognormal"}

\item{cov_vars}{if \code{cov_dist_opt} one of \code{c("MVN", "AFT_lognormal")}, a list of character strings indicating the names of the variables from \code{data} to be used as predictors in the covariate distribution Otherwise \code{NULL}}

\item{cov_mean_user}{if \code{cov_dis_opt = "user MVN"}, the mean of the multivariate normal distribution of \code{(log(X), log(C), Z)}}

\item{cov_sigma_user}{if \code{cov_dis_opt = "user MVN"}, the covariance matrix of the multivariate normal distributionof \code{(log(X), log(C), Z)}}

\item{gh}{if \code{TRUE} (default), gauss-hermite quadrature will be run to estimate the integrals. Otherwise, the \code{integrate} function will be used.}

\item{gh_nodes}{number of nodes to use in gauss-hermite quadrature.}

\item{...}{additional arguments passed to function \code{multiroot}.}
}
\value{
A list with the following elements:
\item{beta_est}{a vector of the parameter estimates.}
\item{se_est}{if \code{sandwich_se = TRUE}, a vector of standard error estimates from the empirical sandwich estimator. Otherwise, \code{NULL}}
\item{iteration_count}{the number of iterations used in \code{multiroot}.}
}
\description{
Performs an augmented inverse probability weighting estimation for a regression model with censored covariates.
}
