% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_aft_acc}
\alias{weights_aft_acc}
\title{AFT Weighting for ACC}
\usage{
weights_aft_acc(data, cens_ind, weights_cov, Y, cens_name)
}
\arguments{
\item{data}{a data frame containing columns for the censoring indicator, the outcome, and the all covariates involved in the weights}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{weights_cov}{a list of character strings indicating the names of the variables from \code{data} to be used as predictors in the weights model}

\item{Y}{a character string indicating the name of the outcome from \code{data}}

\item{cens_name}{a character string indicating the name of censored covariate from \code{data}}
}
\value{
A vector of weights
}
\description{
Performs log-normal accelerated failure time regression models to find the inverse probability of being censored given fully observed covariates.
}
