% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_cox}
\alias{weights_cox}
\title{Cox Weighting for IPW/AIPW}
\usage{
weights_cox(data, cens_ind, weights_cov, cens_name)
}
\arguments{
\item{data}{a data frame containing columns for the censoring indicator and the all covariates involved in the weights}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{weights_cov}{a list of character strings indicating the names of the variables from \code{data} to be used as predictors in the weights model}

\item{cens_name}{a character string indicating the name of censored covariate from \code{data}}
}
\value{
A vector of weights
}
\description{
Performs a cox proportional hazards regression model to find the inverse probability of being censored given fully observed covariates.
}
