% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_mvn}
\alias{weights_mvn}
\title{Multivariate Normal Weighting for IPW/AIPW}
\usage{
weights_mvn(data, cens_ind, weights_cov, cens_name)
}
\arguments{
\item{data}{a data frame containing columns for the censoring indicator and the all covariates involved in the weights}

\item{cens_ind}{a character string indicating the name of censoring indicator from \code{data}, defined to be \code{=1} if observation is uncensored and \code{=0} if observation is censored}

\item{weights_cov}{a list of character string indicating the names of the variables from \code{data} to be used as a predictor in the weights model}

\item{cens_name}{a character string indicating the name of censored covariate from \code{data}}
}
\value{
A list with the following elements:
\item{weights}{a vector of weights}
\item{mu_joint}{the mean of the joint multivariate normal distribution of the log of the censored covariate, the log of the censored value, and the fully observed covariate.}
\item{Sigma_joint}{the covariance matrix of the joint multivariate normal distribution of the log of the censored covariate, the log of the censored value, and the fully observed covariate.}
}
\description{
Calculates the inverse probability of being censored given one fully observed covariate and the censoring value by maximizing the likelihood of a multivariate normal. The multivariate nornmal is the joint distribution of the joint multivariate normal distribution of the log of the censored covariate, the log of the censored value, and the fully observed covariate.
}
